examples/python-basic-generate/requirements.txt
renderbase>=1.0.0
python-dotenv>=1.0.0
